Duplicate User and Cookie
Updated the Passport.js to check that if the user already exists in mongoose then no new user is created.  L26 to L35.

L8 to L16 - is referencing each to the mongoose id - serializing and then deserializing.

Updated index.js for the cookie's library (L3);

L12 to L17 - creating cooking and putting it to last for 30 days (30 days * 24 hours * 60 mins * 60 sec * 1000 millisecs).